Fix Player Login/Logout issues
